python/packages/kagent-core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kagent-core"
version = "0.1.0"
description = "kagent common library for kagent python packages"
readme = "README.md"
requires-python = ">=3.12.11"
dependencies = [
  "a2a-sdk>=0.2.16",
  "opentelemetry-api>=1.36.0",
  "opentelemetry-sdk>=1.36.0",
  "opentelemetry-exporter-otlp-proto-grpc>=1.36.0",
  "opentelemetry-instrumentation-openai>= 0.44.3",
  "opentelemetry-instrumentation-anthropic>=0.44.0",
  "opentelemetry-instrumentation-httpx >= 0.52.0",
  "opentelemetry-instrumentation-fastapi>=0.52.0",
  "opentelemetry-instrumentation-google-generativeai>=0.47.3",
]

[project.optional-dependencies]
test = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=4.1.0",
]

[tool.ruff]
extend = "../../pyproject.toml"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.hatch.build.targets.wheel]
packages = ["src/kagent"]